refactor(js): move state managers into entries/managers
