+++ mason/search/cross/cross_search_using_female.mas
@@ -159,7 +159,7 @@ jQuery(document).ready(function (){
 
     });
 
-    const searchManager = window.jsMod['search_state'].create({
+    const searchManager = window.jsMod['managers.search_state'].create({
         submitButtonSelector: ['#search_crosses_female_male', '#search_all_crosses_using_female'],
         elements: {
             female_parent: { selector: '#cross_female_parent' },
