+++ mason/search/cross/progeny_search_using_female.mas
@@ -158,7 +158,7 @@ jQuery(document).ready(function (){
     });
 
 
-    const searchManager = window.jsMod['search_state'].create({
+    const searchManager = window.jsMod['managers.search_state'].create({
         submitButtonSelector: ['#search_pedigree_female_male', '#search_all_progenies_using_female'],
         elements: {
             female_parent: { selector: '#pedigree_female_parent' },
