+++ mason/search/cross/progeny_search_using_male.mas
@@ -158,7 +158,7 @@ jQuery(document).ready(function (){
     });
 
 
-    const searchManager = window.jsMod['search_state'].create({
+    const searchManager = window.jsMod['managers.search_state'].create({
         submitButtonSelector: ['#search_pedigree_male_female', '#search_all_progenies_using_male'],
         elements: {
             male_parent: { selector: '#male_parent' },
