+++ mason/search/people.mas
@@ -88,7 +88,7 @@ jQuery(document).ready(function () {
 
    var table;
 
-   const searchManager = window.jsMod['search_state'].create({
+   const searchManager = window.jsMod['managers.search_state'].create({
        submitButtonSelector: '#submit_people_search',
        resetButtonSelector: '#reset_people_search',
        elements: {
