+++ mason/search/traits.mas
@@ -138,7 +138,7 @@ jQuery(document).ready(function () {
 
     var trait_table;
 
-    const searchManager = window.jsMod['search_state'].create({
+    const searchManager = window.jsMod['managers.search_state'].create({
         submitButtonSelector: '#submit_trait_search',
         resetButtonSelector: '#reset_trait_search',
         elements: {
