+++ mason/search/treatments.mas
@@ -119,7 +119,7 @@ jQuery(document).ready(function () {
 
     var treatment_table;
 
-    const searchManager = window.jsMod['search_state'].create({
+    const searchManager = window.jsMod['managers.search_state'].create({
         submitButtonSelector: '#submit_treatment_search',
         resetButtonSelector: '#reset_treatment_search',
         elements: {
